csv-handler/src/main/java/com/example/demo/service/DateUtils.java
package com.example.demo.service;

import java.time.LocalDate;
import java.time.format.DateTimeFormatter;

public class DateUtils {

    public static LocalDate parseDate(String dateString) {
        if (dateString != null && !dateString.isEmpty()) {
            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd-MM-yyyy");
            return LocalDate.parse(dateString, formatter);
        }
        return null;
    }

    public static String formatDate(LocalDate date) {
        if (date != null) {
            DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern("dd-MM-yyyy");
            return date.format(dateFormatter);
        }
        return "";
    }
}

